services/file-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
psycopg2-binary
python-jose[cryptography]
